Next and hold screens implemented
Next and hold screens implemented.

The next screen displays the next block the player will be given.

When the player holds a block, it gets sent to the hold screen. If there is already a block in the hold screen, the blocks are swapped. You can only hold a block once per turn.